--- img/0.xpm
@@ -1,73 +1,0 @@
-/* XPM */
-static char *_9174_1aef6cfbd05cb5a6375e51420d12625f0626d7a3f371ce69124101c4975697ed[] = {
-/* columns rows colors chars-per-pixel */
-"40 40 19 1 ",
-"  c #C6AB83",
-". c #CBAD85",
-"X c #CAAF89",
-"o c #D0AF85",
-"O c #CEB087",
-"+ c #CEB28C",
-"@ c #D0B186",
-"# c #D2B58D",
-"$ c #D8B78E",
-"% c #D6B88F",
-"& c #D8B88F",
-"* c #CEB690",
-"= c #D2B791",
-"- c #D5BA93",
-"; c #D8BC94",
-": c #D6BD99",
-"> c #D8BE99",
-", c #D6C09D",
-"< c #DAC09B",
-/* pixels */
-"*+.*.==X*-,-*,-,-*,,<+<+<<O< O,,O,O <,,., ,   XX",
-"*+XX.#==*:-:*:-<----+<<<<+<O<+<,,,,,,O,,O,, XXXX",
-"XXX X#==*,---,-:,-,-<,+<<<+<O,,+,<O,,,,O,O OX+++",
-"*+XX.#--*,--:**-----<@<<<<<<,<<,O,,,O, ,O,X,++OX",
-"#+.OX#--*,<*--=---##<#<#<+<<,O,<,<,<< O,,  O, OX",
-"+=X#==-:--,-,<%%--#<#<<<<<<O<<<,,,,O,,<OO,,OXOO,",
-"+XO=----=,<--,-<-#,<##<$<<<<<<O<<@O,O<O,,,OOOO,O",
-"=OX==-*-,=,=,--%-###%<$<<<<O<O<,O,<<@O,,OO, , OO",
-"=++##%--::---<-#@<@<$<$<<@<<@@<<O<O.<O,@,O<@+O+O",
-"+=##===---==--%<<@<$$@$@<<<<<<+<,<<,,@@,.,O ,O<,",
-"==##=&=----=%---@<@<<$$<<<<@<<<<<OO,<@@@,OO,O<<+",
-"=+=#==---;;&<;--$<<$$<<$<<<<<<<O<O<OO<,,<O< ,.+*",
-"=+=*=--%;;;&->>>$<<<@<<<<<<<O<+<,<O<@O<O,,.,+**O",
-"-=+=#=;-;#;#--;;<@@@<<@<<<@<<<<O@<@<<<@<O<, <++<",
-"-==-==---=;--->;@<<@<@$$@<<@<<<@O@<@@,< <@O,#,##",
-"=+-==---;-;;>>--$$$<<@<<@<<<@@<<<<@<@<<.<O ,##O.",
-"====:--::==&;--;<@$$@<@<<@<$<<<$;;---=%--###+OOO",
-"==-=-:-:;:&;;--;$<@<<@<$$<$<<$$<--;-#--#+#@O####",
-"===::---;-;--::;@@<$$<$<$$$<<<<<:;;-;-=$#O%O*=#*",
-":====::----&--&&<<<@<@<$<<<@#$<#;;-:-;#$$#%+*-#-",
-"===:-%%--&&-&-&&@<<@<@<$%#<#<<$<-;;;;==##O+*%#O*",
-"*=-==-;;-#--=&=%@@$$<$$@&,%<#<<#:&&&=;==%#%**+**",
-":=-==-:-&--&%$$%$$$@$$$$&<<#<<<<;;--=-:--*******",
-"*==:-:-;=-&;&=&&$$$<$<$<%&<#@#$<;;;-=::=-%*%#*-*",
-"==-=&=&-&=$&=%-=$%;&&--&&<%%:>%>>-&$#::-===****+",
-"=*=#==&==$#&&%&-$%&&;&-------;;;&&;$=:-:::*+*X+X",
-"===#--&====-&=&&&#&&&&-&-%-:$;;%&&##&--:==+*=+++",
-"*=*==--=$==-=&=&=&-&=;-&$$-%%#%%&&o$=;:-==:***++",
-"++=#=#=--=&--:-&&-&&&=;&=&=--$-$$$$$=::;:==#=XX+",
-"#===--::-=:---=-#&--;=--=#;-;---=$o&;;:==#+==X++",
-"==----:-=-:&---===---:-;&#$;-->#-=#:;>=$####X*XX",
-"====--:===--:-=-#=-&:-;-=;;=-%--:--=--+$##===*==",
-"====:=:====:=:=-=>$=>==>$=:=>$>=:>==:$=====+=+X+",
-"====:-**=-:--=--=:==>==>>==>=>=>>=:$=$====+$++*+",
-"=====:====:=::-=:==:=>:>>==>===>;==>=:$====+*++X",
-"======-=:*=:====$=======$==>:>==:>=>:=$====$X+XX",
-"*=====---*=-==#===$$=:$=:=>=:=>>=>;:=:==o=o++XX+",
-"=*=====*++==============:====>=:=:::===oo$o=+ooX",
-"==*======X*==###=$===$+===>>>>==:======$=oo=+oX ",
-"+=+**=*==========:$======$=$$=>==;;==$o=$oo#+++X",
-"====$++=*=====++$==$====+=$$====;==$$=o#o+o++X X",
-"====++++XX====##$=============$=-====$o++#+o+o.+",
-"X==X$X++*X+=#o=o+o=$==+==$$=$$=$=$###+ooo++o.XX+",
-"=X++==++X+=+#++==X$o==+$=+====$=##=o#=ooooo...XX",
-"+++==+++oo=o####o+$+=$$+=+$=$==+=o#++oo+o.o. o  ",
-"++++=+++=+++#+##+===++=+===++====+#=#o++o.X.    ",
-"+X+=+X+XX+++=##=$==+o$+$$====+=++##o+o#+o.XXX . ",
-"+++++o+=o+$+++##=$+==$++=+==$+++#oo#+oooXo  ...."
-};
