--- img/1.xpm
@@ -1,159 +1,0 @@
-/* XPM */
-static char *c30d69bd32264a19aaf9192b5986959dEgzSAJkuubbOSdAh[] = {
-/* columns rows colors chars-per-pixel */
-"40 40 111 2 ",
-"   c None",
-".  c #2F2C2E",
-"X  c #312D2F",
-"o  c #322F31",
-"O  c #353233",
-"+  c #383536",
-"@  c #3F3C3E",
-"#  c #4E4C4D",
-"$  c #514D4E",
-"%  c #565354",
-"&  c #595658",
-"*  c #5C5A5B",
-"=  c #5F6065",
-"-  c #666365",
-";  c #696667",
-":  c #65676B",
-">  c #696A6D",
-",  c #726C6C",
-"<  c #7A6D6B",
-"1  c #6F7175",
-"2  c #727070",
-"3  c #727478",
-"4  c #BD2927",
-"5  c #DB3B08",
-"6  c #D64715",
-"7  c #DA4714",
-"8  c #DA4815",
-"9  c #DC4F1C",
-"0  c #DD511C",
-"q  c #DD5422",
-"w  c #C7623F",
-"e  c #D16035",
-"r  c #A76564",
-"t  c #A76869",
-"y  c #A96D73",
-"u  c #B26D73",
-"i  c #B37777",
-"p  c #B57979",
-"a  c #EB7B56",
-"s  c #EB7E59",
-"d  c #EB815C",
-"f  c #E59173",
-"g  c #7A7E83",
-"h  c #827F80",
-"j  c #7B8085",
-"k  c #7E8388",
-"l  c #868384",
-"z  c #888687",
-"x  c #81868B",
-"c  c #898788",
-"v  c #84898F",
-"b  c #8A8889",
-"n  c #968D8B",
-"m  c #858C92",
-"M  c #888F96",
-"N  c #8A9095",
-"B  c #8D949B",
-"V  c #9A9798",
-"C  c #9C9A9B",
-"Z  c #BE8A8E",
-"A  c #BF8D92",
-"S  c #A19E9F",
-"D  c #959DA4",
-"F  c #9A9EA4",
-"G  c #A09FA0",
-"H  c #99A1A9",
-"J  c #A5A4A4",
-"K  c #A8A7A7",
-"L  c #A2A5AA",
-"P  c #ADABAC",
-"I  c #B8A7AC",
-"U  c #B2ADAD",
-"Y  c #B1AFB0",
-"T  c #B2B2B2",
-"R  c #E39980",
-"E  c #C1B1B0",
-"W  c #C6B9B7",
-"Q  c #C0B6BC",
-"!  c #B5BDC3",
-"~  c #BAC2C7",
-"^  c #C6C5C5",
-"/  c #C8C7C7",
-"(  c #CAC9C9",
-")  c #D8C7C4",
-"_  c #DCD0CC",
-"`  c #D6D5D5",
-"'  c #E0DFDF",
-"]  c #DEE2E5",
-"[  c #E1E1E1",
-"{  c #E8E7E7",
-"}  c #EEEEED",
-"|  c #F3E8E4",
-" . c #F0EFEF",
-".. c #F0F0EF",
-"X. c #FBF2EF",
-"o. c #F1F1F0",
-"O. c #F2F1F1",
-"+. c #F3F3F2",
-"@. c #F4F4F3",
-"#. c #F7FAFB",
-"$. c #F9F9F8",
-"%. c #FAFAF9",
-"&. c #FBFAF9",
-"*. c #FEFCFB",
-"=. c #FFFDFB",
-"-. c #FFFDFC",
-";. c #FFFEFC",
-":. c #FEFEFD",
-">. c #FCFDFE",
-",. c #FFFFFE",
-"<. c white",
-/* pixels */
-"                                                                                    ",
-"                                                                      X   X         ",
-"    O - z z l l z b b b l l h l l l l c l l l c l l l l b c z b z z l l z b & O     ",
-"    O K <.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.b o     ",
-"    o J <.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.b o     ",
-"    o J <.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.b o     ",
-"    O J <.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.c O     ",
-"    o K <.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.z o     ",
-"    o K <.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.<.z o     ",
-"    o J <.<.<.<.<.<.<.<.( V C S C C V V C C C C C C C V S ` <.<.<.<.<.<.<.<.l X     ",
-"    O J <.<.<.<.<.<.<.{ % ; l l h l l h l l l l l l l l ; - @.<.<.<.<.<.<.<.l .     ",
-"    O J <.<.<.<.<.<.<.' # / <.<.<.<.<.<.<.<.<.<.<.<.<.<.Y % } <.<.<.<.<.<.<.c X     ",
-"    o J <.<.<.<.<.<.<.[ # ^ <.<.<.<.<.<.<.<.<.<.<.<.<.<.P & ..<.<.<.<.<.<.<.l o     ",
-"    O K <.<.<.<.<.<.<.[ # ^ <.<.<.<.<.<.<.<.<.<.<.<.<.<.P * o.<.<.<.<.<.<.<.l O     ",
-"    O K <.<.<.<.<.<.<.[ # ^ <.<.<.<.<.<.<.<.<.<.<.<.<.<.K * o.<.<.<.<.<.<.<.b O     ",
-"    O K <.<.<.<.<.<.<.[ # ^ <.<.<.<.<.<.<.<.<.<.<.<.<.<.K &  .<.<.<.<.<.<.<.c o     ",
-"    O K <.<.<.<.<.<.<.[ # / <.<.<.<.<.<.<.<.<.<.<.<.<.<.U * ..<.<.<.<.<.<.<.c o     ",
-"    O K <.<.<.<.<.<.<.' # ^ <.<.<.<.<.<.<.<.<.<.<.<.<.<.P * o.<.<.<.<.<.<.<.b o     ",
-"    O K <.<.<.<.<.<.<.[ # ( <.<.<.<.<.=.<.<.<.<.=.=.=.<.P * o.<.<.<.<.<.<.<.c o     ",
-"    O J <.<.<.<.<.<.<.[ # n f s d d s s d d d s a a a R l * o.<.<.<.<.<.<.<.c o     ",
-"    O J <.<.<.<.<.<.<.%.G < e 0 q q 7 5 6 6 5 8 9 q 9 w , P <.<.<.<.<.<.<.<.l X     ",
-"    o K <.<.<.<.<.<.<.<.<.%.| E I _ p 4 r t 4 i I W ) X.%.<.<.<.<.<.<.<.<.<.c X     ",
-"    o K <.<.<.<.<.<.<.<.<.<.<.] ! #.Q y Z A u I ! ~ <.<.<.<.<.<.<.<.<.<.<.<.b X     ",
-"    O 2 P K P P P P T U P P U K F T F D K K F L L C Y Y P P Y P P P P T T P ; o     ",
-"      O O O O O O O + + O O O X = 1 : m > = m k g @ + + + + + + + O O + O O O       ",
-"                                  D D * j m 3 D D                                   ",
-"                                  H D x 3 > B D F                                   ",
-"                                  D   M 3 3 x D D                                   ",
-"                                  D D > v x g D D                                   ",
-"                                  H D B N m B D                                     ",
-"                                      1 * * -                                       ",
-"                                      $ # #                                         ",
-"                                      $ # #                                         ",
-"                                      $ # #                                         ",
-"                                      $ # #                                         ",
-"                                      $ # #                                         ",
-"                                      $ # #                                         ",
-"                                      $ # #                                         ",
-"                                      $ # #                                         ",
-"                                      $ # #                                         ",
-"                                                                                    ",
-"                                                                                    "
-};
