+++ img/eagle.xpm
@@ -1,0 +1,126 @@
+/* XPM */
+static char *_832aa764d464016d1343c840f912677[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 56 1 ",
+"  c #202020",
+". c #2C2C2C",
+"X c gray19",
+"o c gray22",
+"O c #400000",
+"+ c #4C0000",
+"@ c #580000",
+"# c #640000",
+"$ c #700000",
+"% c #7C0000",
+"& c #543C1C",
+"* c #585400",
+"= c #706C00",
+"- c #744C28",
+"; c #400040",
+": c #480048",
+"> c #500050",
+", c #580058",
+"< c #600064",
+"1 c #6C0070",
+"2 c gray25",
+"3 c #484848",
+"4 c gray33",
+"5 c #646464",
+"6 c gray44",
+"7 c #7C7C7C",
+"8 c #880000",
+"9 c #980000",
+"0 c #A40000",
+"q c #B00000",
+"w c #BC0000",
+"e c #C80000",
+"r c #D40000",
+"t c #E00000",
+"y c #EC0000",
+"u c #FC0000",
+"i c #FC2020",
+"p c #B45400",
+"a c #CC6000",
+"s c #FC7800",
+"d c #A86840",
+"f c #848400",
+"g c #9C9C00",
+"h c #FC8820",
+"j c #CCC400",
+"k c #E4D800",
+"l c #FCF400",
+"z c #E88C58",
+"x c #840084",
+"c c #9C009C",
+"v c gray55",
+"b c #989898",
+"n c gray66",
+"m c #B4B4B4",
+"M c #C0C0C0",
+"N c #D0D0D0",
+/* pixels */
+"oooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooo",
+"4o433333433333343o333333344444433333333344444443333ooo334433333o",
+"oyyrrrrrrrrr0334yyrrrrrrrrrrrrrrrrrr03o4yyyyttttttrrrrreeeewww03",
+"3leqqr80r8e0#.34yw8w8w8w%w9w9w%w%we0$.3or0eqeqqe$e%eqe8q0%e9e0l.",
+"kjfmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmjj=",
+"dg*6677777777777777777777777777777777777777777777777777777666=z*",
+"3&07777bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb7677@d.",
+"3w%b7i0OO4$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$0qut7OOO7M6@O.",
+"3w0n70Mb6O7nMNMMmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmMMMMM7O76bi76@O.",
+"3%#mbObMb7vbnnv7nNMmmmmmmmmmmmmmmmmmmmmmmmmmmmmv15Mmnnn6nN067OO.",
+"33.NnObnMb77bb4,6nNNMMmmmmmmmmmmmmmmmmmmmmmmMMv5<>bnnbbnN7O7b  .",
+"4ooNm46bmMb77v>,,,7nNMMMMMmm>64>>>>pasmmmmMMv5<,>>bbbbnNM54vn ..",
+"yrrmNO7bnmMb653:>>,,7nmmmmmm>>>>>7>ppahmmmv5<,>>>3bbbnMMn7Obn@@r",
+"q88mN$bvbnm5,653::>>,,,7vmmm6>>>>>>pvmmv5<<,,>>>366>5nMnv7Onm@@q",
+"808mNqnbvb51<3653::::x,,,6vmm>>>>>>vmv5<<,x>>>:3641<<5nb7v$nm@@9",
+"$00mm9mbvv:<1<,353o::1:,,,,mm>>>>>>mm,<<,>1::o353><<11bv7v$nm@@%",
+"088mm8Mnbv::<1<<,34o;;:::,4mm>>>>>>mm4::::::o33:>,<111v7vv$nm@@%",
+"$0#mm$Nmnb3::<11<,:;;;;::vmm6>>>>>>6mmv:::;;;::>,<1114v7vb$mm@@8",
+"$q#mm$NMmb53:,<<1<,1;;;:v>6m>>>>>>>>m6:v:;;:,::,<111477vvn$mm@@$",
+"O++mm$NN5>6443,<<1,x>;;:>:>mm>>>>>>mm::::::,1>,<<1475335bn$mmO@O",
+"...mm$mN,>:34443<<1,x>>>:>:75>>>>>>57::::::1,,<<4v753;;:bn$mm  .",
+"443mm$mN<,>::;o3345,,,,:::>>>>>>>>>>::::::>>,45v74>:;;;:nn$mm ..",
+"4yyMM%Mm<<>>:;;;::,c,,,::>>>>>>>>>>>:::::>,,c,,,>>:;;;;,nn$mm@@.",
+"4y8NN9NM,<,>::;;;::x,,,>>>>>>>>>>>>>>::>>,,,c,,,>::;;;:,bn%mm@O.",
+"oyqMM8Mb6,<>>:;;:;:1,,>>>>>>44566544>>>,,,,,x,,>::;;;;:6bn%mm@O.",
+"3w#mm%mbb63,>:::>:>,1>>>>>>>51c11c15>>,,,,,x,,>:::;;;36bbm8mm@O.",
+"3wqmm$m>,356655432X:>,x,>>>>7c1111c7>,>,,x,,,4432o23453<1M9MM@O.",
+"4w$nn$n>>,,,,>:>:>:>:>,x1>>>41111114>>,1x,,,,>::::;::>,<1N9NN@O.",
+"3e0nn#n>>>,,,>>:>:>:>>>>>>>>7c1111c7>,>,>,,,>:::::::>,,<<M9MM@O.",
+"309bb+b>>,>,>,>>>>>>>>>>>>>>41c31c14>>,>,,,>>:::::::>><1<M8MM@O.",
+"33onn@n>>>,>,>>>>>>4>>>>4>>>74131147>,>4,>>>4:::::::>><<>m%mm  .",
+"444nn#n5>>>>>>>>36>>>>>4>>>>47744774>>>,4>>:>:63:::>>,<,5m$mm ..",
+"wwwmm$mb5>>>>366>>>>>>6>>>>>>>4774>>>>,>>6>>:::>663>>,,5bm$mm@@w",
+"0q0mm$mmmbbbb6>>>>>>>6>>>>>>>>>>>>>>>>>>>>6:>::::>6bbbbmnn#nn@@0",
+"0%8mm$mmbb6>>>>>>>>>6>>>>>>>>>>>>>>>>>>>>>>6:>:>:>,,,6bbbb@bb@@$",
+"8%0mm$mm>>>>>>>>>>>6>>>>6>6>>>>>>>>>>6>6>>>>6:>:>,,,,,>>vv+vv@@$",
+"08%mm$mm>>>>>>>>>>6>>>>5b6>>>>>>>>>>>>6b5>>>>6>>>>,,,,>>bb@bb@@8",
+"08%mm%mm6>>>>>>>>6>>>>>b>>>>>>>>>>>>>>>>b>>>>>6>>,,,,,>6bb#bb@@$",
+"$$8mm%mmb>>>>>>>6>>>>>b>>>>>>>>>>>>>>>>>>b>>>>>6>>,>,>>bnn#nn@@$",
+"OO+mm8mmm6>>>>>6>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>6>>>,>6mmm#mm@@O",
+"...MM8MMMm6>>>6>>>>>>b>>>>>>>>>>>>>>>>>>>>b>>>>>>6>>>6mmmm$mm  .",
+"434NN0NNNmmbbbb>>>>>bmc>>>>>>6>>>>6>>>>>>cmb>>>>>bbbbmmmmm$mm ..",
+"oyrMM9MMMmmmmmbb>>>bmmm6>>>>>>>>>>>>>>>>6mmmb>>>bbmmmmmmMM8MM@@.",
+"4r0mm9mmmmmmm7pp7mmmm7p>>>>>>>>>>>>>>>>>>p7mmmm7pp7mmMMMMM8MM@O.",
+"3r0mm8mmmmmmmpppppppppp-->>6m>>>>>>m6>>--ppppppppppMMMMNNN9NN@O.",
+"3r8mm8mmmmmmmm75-pppp--7mmmm>>>>>>>>mmmm7--pppp-57mmmMMMMM9MM@O.",
+"or0mm%mmmmmmmvppppppppppmmm>>>>>>>>>>mmmppppppppppvmmmmmMM8MM@O.",
+"4r0nn$nnnnmmmp--vpp-mv-pm7>>>>>>>>>>>>7mp-vm-ppv--pmmmmmmm$mm@O.",
+"oe0nn$nnnnnnm-7vmp-mmmm74>>>>>>>>>>>>>>47mmmm-pmv7-mmmmmmm$mm@O.",
+"40#mm$mmmmmmmmmmmp-mm74>>>>>>>>>>>>>>>>>>47mm-pmmmmmmmmmmm$mmOO.",
+"o3.mm$bmmmmmmmmmmvpm>>>>>>>>>7>>>>7>>>>>>>>>mpvmmmmmmmmmmb$mm  .",
+"4o4mm$7mmmmmmmmmmmmm74>>>>>5m>>>>>>m5>>>>>47mmmmmmmmmmmmm7$mm ..",
+"rrrmm7$7bmmmmmmmmmmmmm74>>5m7>>>>>>7m5>>47mmmmmmmmmmmmmb7$7mm@@w",
+"%0@7mm7$$$7bmmmmmmmmmmmmmmmm>>>>>>>>mmmmmmmmmmmmmmmmbv$$$7bm7@@$",
+"0%%@7mmmb7$$$$7bmmmmmmmmmmmm5>>>>>>5mmmmmmmmmmmmb7$$$$vbmmm7@@@+",
+"0%@@@47mmmmmb7$$$7bmmmmmmmmmm5>>>>5mmmmmmmmmmb7$$$7bmmmmm74@@@$8",
+"%9@@@@@47mmmmmmb7$$$7bmmmmmmmm5>>5mmmmmmmmb7$$$7bmmmmm74@@@@@#$0",
+"#0%0#@#@@@@47mmmmmb7$$$7bmmmmmm55mmmmmmb7$$$7bmmmmm74@@@@@@$@0$0",
+"#0%%#99#@@@@@@47mmmmmb7$$$$7bmmmmmmb7$$$$7bmmmmm74@@@@@@@$0$@#0+",
+"%9@0##8#%%0@@@@@@O47mmmmmb7$$$7bb7$$$7bmmmmm74O...@@@@0@0$0$0#0$",
+"++++++++++++++++OO  .47mmmmmb7$$$$7bmmmm74+++OO ..@OO+++++++++++",
+"...................     .47mmmmmmmmmm74     .   .........     ..",
+"o4o4ooo333334oooo.......     47mm74.     ........oo33433332o2333",
+"44444444444444444443ooo..                 .ooo334444444444444444"
+};
