+++ img/greystone.xpm
@@ -1,0 +1,92 @@
+/* XPM */
+static char *_9bbf054b48d4672a205434ff8068a49[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 22 1 ",
+"  c gray12",
+". c gray23",
+"X c #3F3F3F",
+"o c gray28",
+"O c #4B4B4B",
+"+ c gray31",
+"@ c #535353",
+"# c gray34",
+"$ c #5F5F5F",
+"% c #676767",
+"& c gray42",
+"* c gray45",
+"= c #7B7B7B",
+"- c #838383",
+"; c gray56",
+": c #9B9B9B",
+"> c gray67",
+", c #B7B7B7",
+"< c gray78",
+"1 c LightGray",
+"2 c #DFDFDF",
+"3 c #EFEFEF",
+/* pixels */
+" . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .",
+"#@#+#+#+#+O1222121222121222 .->,1<11<1,<>-# oooooooooo@+#+#+#+#+",
+"@#@@#@@@#@o21<1,111<11,11<; .-<,;>-;;;;->+. .&;;-;;;=&%oooOoooOo",
+";;-;;;-; @O2<<,1,<<,<<,,1;$ .>1>;>;>->;>-#. .;122212,,>;-;;;-;;;",
+"1111,>;- #o21>>:>;>>>>:>>=@ .;,;;->>>>-;*+. .-22,<,1222121<111<1",
+">>>>:>=* @o3,:>>>>:>;>->>&@ .o>;>>-;:;>&#@. .;1,>;>;>>:;;>;>->;:",
+",:::>;=& Oo31>->;>;:::;;>;+ .o=;->;>>->*... .-2,;>-:-;;>-;:;>:;:",
+",>-,-:$# .o3<;:;>>:;=;;>;$# .oo;>>:;;;>&... .;2>>-:=>:=;>->>;=>:",
+":>>;:*oO  O<,>:>->;:-:;;>&@  O@o=........  ..*22;>--:*:;>;;=>>*;",
+";;-;*@oo .o>3>:>;;;;;;->-#=  .o....... .   ooo<<;;;;:::-;>;>>->;",
+"*&**+O..  O;1;>;;=;;>->;>&#. .o.       . . O@o<>;>->;-;*&%$$%$%$",
+"@#ooo    .X*2>-:===>:;-=;&%. .ooOO*=;>:>>:*.+O,;&*%$%#@@@oOoOooO",
+"     . . .@o2<;;;::-;>=;>;+. .X*>>123111:&@ @o;&%$@Ooo..... . . ",
+" . .  . ..+o>1;>-=;>;-:;-*+.  .>,21,->;>-*. ...........         ",
+"..........OO;<>->:-=;=;>;+#. .X>2,>;;;-;;+. .  .  .  . XX.......",
+"@O...... .@o*<->:;>:>->->+*o  O>3>;->;>->@. .o#@## oOooOoooooooX",
+"::;:;:;::  #o<,:;-::*;;;;&#@  O>3-;>>->;;+. .&>21#o@=::;:::;::::",
+"1<11<1,>*X #o;,;>;=;;->;&;&O  O><>>;->;>-#. .>1<> o=<12221222121",
+">:;>;>>&*. @o*:>->>->*;&@*@O  O><>;&**&@#@. .1<>- O:11,,>,,,>,,,",
+">;:;>-;&*. o@o:;**&*@@*+*@o   .:;*;&....... $2>;; o:2>>:>;>;:-::",
+">;;>->>&#. o@o%@#@#ooo.        ;*...        $2>>- O:1>;:-:;:*-;-",
+">;>;>-;*+. o@o . .            ..+.     .  . ;2>>; o;<->:>;>;:-::",
+";>*=;;>&#. .+#O.        +OO+OOo;;;>>;;;=# X%,2>-= O-<>-:;:-:==;;",
+";--=;>-%o. .-2221<>>;$.O%=>>,<<2221222<:% .=2<:>* o-,>:;>->;==;;",
+";;>>-;;$O. .1<>;::>;:.+O=1322<,,<,<,<,<&# .;2>:;& O&<->;;:;->;>;",
+":>-;>-*@.. .2:;:>>;:> @O:2<;>->:->:>:;-%o .;3>-># Oo;*$#@@%%&=;-",
+"%$%$%@o... .1>>>->;>$ Oo:3>:>;=>>;-;>->#o .-3>>;# @ ..ooooOoo@#@",
+"@Oo...     .2->;>>;=$ Oo:3->;:>-;=;>;;;+O .;3->;+ #.         ..o",
+"       . OOo2>:->;>=$ .o:3;;>-;=;=;-;>;+O .-3>>-# #....... .    ",
+"o%%&;:;-;;o 2>*:;>--%. o;3>;-;=;=;>=;->@o .;3;>>+ #oXo+@+#@@#...",
+"%;,21211<>* 2>>-=;>-%. o;2,:;>;>*=;=;>-#. .;1;;>+ O>21,;$#ooOoo@",
+"%>1<<,>>:>X 3;:>=;>:&o O=1<>-;*;==>;>--+. .;2>;;@ O2-;,221222;.@",
+"%2,>>->;>-* 3:>;>>->&O .+<2>>:;>->==;>&#. X;1;>;+ O2;->>-;>;--.+",
+"%1,:->;>->* 3:-:>;>-%@  #>3>>:;;=;==>-*o. O;1>>-# o2>>-;>>->;$.o",
+"%2>;>;>>=>+ 2>>>;;;>%+ .+;3,>;>-;:->;>$o .X;1:;>+ O<>;>>-;>;>$.o",
+"%2>>;;::>-# 2,=:=>>:$#  O&1,>;=:-;:;;;$.  O;1;>+# O<>:>;>-;>:#.o",
+"%1,;>>;>->@ <2-:>;;;$#  O@,2>->;>;>-;%@. .o=>=+...X>>;>:;>;->#..",
+"%2,>:;>;>;+ ,2=>-;;>&#  .+;,>;=%$@oO.... .X#....Xo *+o@#@#@@#@..",
+"%2>;>>->=>@ :1:=>>-;&#  .@o;;$#o.        .o+oo+@#@#            .",
+"%1,;-:>;>-# =2>:>-;>%+   +o. .    . .  . .o@#ooOoooOoo..........",
+"%2,>>-:>;>@ #3;:;;>-%@  .ooo.....OoOX+OoOoo#@================*.+",
+"%2>;;>->;&# o3>;-=;>&#   X%=============;;;:>,><,<,<,,<,<,<,<=.+",
+"%<,>>;>;>+  +3:;>;;-%#.  .=<,<,<,<<,<,<,<,122333333333331211;% O",
+"%1,;>;>>-# .X3;>=:->%+.   =<<3333322332333332211,,,<,>,<,,<,=# +",
+"%,,>;>-;;+ .o2>>->;;$#.   =<<,<,<,<>,<>,<,,>>::;;>;>-;;>->;>&#..",
+"O<-;>-*&*@ .o1-;>->;&#.   %,2,;;-;;;>->>;;:>>>;=:->>:->;>>:;&@ .",
+"o,>>:;#@@  Xo,>-;>;-*@.   o<2,;;;>;:;:>-;:>>-:;:;;=-::;>->;>&#..",
+"o===#@#@@ ..X::>-;;&%@.    ,21>>-;;;=;->;:->:=-;>-;>;=*;=:;;*@ .",
+"o.   #@@# .XO=%$%%$%@@.   .>11;->;>-;>;==::->;>;;>-;=;;->;*-&# .",
+"@#@o.    ..X...........    ;1,>;;>->=;-:=:;:=>-;=;;;->;=;;-=&# .",
+"oOoooOo.  .O..            .-:&**&**&>;>>-;=-;;=;;>->;;>-;;;:&# .",
+"o=>>>>;;%OO@..  .  .  . . .@o.. . . #*&=;===================%+ .",
+"o><12122,>>;&$#O..........X#oo...... O@@#@#@#@#@@@#@#@@#@#@@@ ..",
+"o>1>>;,1222,>>>>>>>:;;-==&%%+#oooXo.. . . . . . .             .X",
+"o>2>>>,>;>,222121212222<1<,>>>>>-O@o.........  . . . . . . ...oO",
+"o>2>->==>:;=>->;>;>>->;>>><1211,-.o+Ooo... .................XoO@",
+"o><>>>;>;;;:>-;>=*;>=;-;>->;>>>=% .+O>12:=OOo@@@ooXoooXooXoo+#@@",
+"o><>->==:>->>:-=>=;=:->;>;>->>:%@ .oo2>>2221,@@O,22122211<>:=Oo@",
+"o>-%$%$%$%%$%$%%$%$%$%%$%$%$%$%#@. .o1,>,>,,,$#=2:>>;>->:>;>1: @",
+"o;%...............................  O,>,,:$:-%o,>;>->>>->;>->@ #",
+"o.                                  +.o.+%$%%$O2:::>>;:>::>;;+ #",
+"@O... .  .  .  . .  .  .  .  .  . . X+#..     .@#@@@@@#+@#@#+# @",
+"@#+@oooOOoOO.........................+#+@......+@              #",
+"................................................................"
+};
