--- img/p.xpm
@@ -1,267 +1,0 @@
-/* XPM */
-static char *_be59d19e4df464ad46b1a6747a5b49d6xBGUSpL1DFOZpNu[] = {
-/* columns rows colors chars-per-pixel */
-"20 20 219 2 ",
-"   c None",
-".  c #0C0A0C",
-"X  c #090607",
-"o  c #130D0C",
-"O  c #190F0E",
-"+  c #1D110D",
-"@  c #13100F",
-"#  c #0C0E11",
-"$  c #110E10",
-"%  c #0D1214",
-"&  c #131416",
-"*  c #1B1515",
-"=  c #121719",
-"-  c #1C1619",
-";  c #15191C",
-":  c #1B1C1C",
-">  c #191916",
-",  c #24160E",
-"<  c #221613",
-"1  c #261A15",
-"2  c #2C1B13",
-"3  c #231D1C",
-"4  c #2A1E1B",
-"5  c #281714",
-"6  c #3C1E1F",
-"7  c #1D201E",
-"8  c #26211E",
-"9  c #2B211D",
-"0  c #352116",
-"q  c #33231B",
-"w  c #3A251C",
-"e  c #3D291C",
-"r  c #3B2317",
-"t  c #1A1D21",
-"y  c #161C21",
-"u  c #251E21",
-"i  c #1D2124",
-"p  c #1F2429",
-"a  c #232324",
-"s  c #2B2322",
-"d  c #26262A",
-"f  c #2E2A2B",
-"g  c #24292A",
-"h  c #332623",
-"j  c #342925",
-"k  c #3B2A24",
-"l  c #352B2B",
-"z  c #3A2E2B",
-"x  c #392523",
-"c  c #3C312D",
-"v  c #2D2B30",
-"b  c #332E32",
-"n  c #3B2F32",
-"m  c #363235",
-"M  c #3C3333",
-"N  c #3C3439",
-"B  c #3D3839",
-"V  c #432C1D",
-"C  c #40261C",
-"Z  c #442D22",
-"A  c #432E2A",
-"S  c #472C27",
-"D  c #4C3224",
-"F  c #43322C",
-"G  c #4C352D",
-"H  c #4D382E",
-"J  c #463127",
-"K  c #513426",
-"L  c #543729",
-"P  c #54392D",
-"I  c #5B3C2F",
-"U  c #433533",
-"Y  c #443935",
-"T  c #4B3A34",
-"R  c #42373A",
-"E  c #453B3C",
-"W  c #4B3D3B",
-"Q  c #493631",
-"!  c #533C34",
-"~  c #593E32",
-"^  c #513E3C",
-"/  c #583F38",
-"(  c #422E32",
-")  c #643E35",
-"_  c #4D403D",
-"`  c #5B4234",
-"'  c #54423C",
-"]  c #5B443B",
-"[  c #5D493D",
-"{  c #534037",
-"}  c #63463A",
-"|  c #674B3B",
-" . c #6A4C3D",
-".. c #644536",
-"X. c #72513F",
-"o. c #453D42",
-"O. c #4A3F41",
-"+. c #533D42",
-"@. c #4B4142",
-"#. c #4D454A",
-"$. c #474046",
-"%. c #534542",
-"&. c #594543",
-"*. c #554946",
-"=. c #5C4B45",
-"-. c #534A4A",
-";. c #5C4C4B",
-":. c #58454A",
-">. c #5E514D",
-",. c #4C4853",
-"<. c #554D54",
-"1. c #5B4F52",
-"2. c #504A58",
-"3. c #5C5354",
-"4. c #59515A",
-"5. c #565354",
-"6. c #624D45",
-"7. c #6B4C41",
-"8. c #624D4C",
-"9. c #6E4F48",
-"0. c #714F44",
-"q. c #6C5244",
-"w. c #63514C",
-"e. c #6A534B",
-"r. c #68544A",
-"t. c #725243",
-"y. c #7A5C4E",
-"u. c #765549",
-"i. c #6B4C53",
-"p. c #624F50",
-"a. c #724F56",
-"s. c #635453",
-"d. c #695655",
-"f. c #6B5954",
-"g. c #635659",
-"h. c #645A5B",
-"j. c #6B5B5B",
-"k. c #6A5859",
-"l. c #725C54",
-"z. c #7C5C53",
-"x. c #725D5C",
-"c. c #775357",
-"v. c #7D6255",
-"b. c #72625C",
-"n. c #7B635C",
-"m. c #776055",
-"M. c #555668",
-"N. c #645D63",
-"B. c #6A5C62",
-"V. c #615B68",
-"C. c #715F60",
-"Z. c #6C6164",
-"A. c #6D646A",
-"S. c #6D686D",
-"D. c #746363",
-"F. c #7B6564",
-"G. c #7A6963",
-"H. c #73646D",
-"J. c #7B666B",
-"K. c #7B6A6D",
-"L. c #726A6F",
-"P. c #6B6772",
-"I. c #726C74",
-"U. c #786B75",
-"Y. c #7D7073",
-"T. c #7F7179",
-"R. c #73707B",
-"E. c #805E4A",
-"W. c #846A5F",
-"Q. c #87665A",
-"!. c #845B63",
-"~. c #836A65",
-"^. c #836B6B",
-"/. c #886866",
-"(. c #916F67",
-"). c #86716B",
-"_. c #887066",
-"`. c #92726B",
-"'. c #98766F",
-"]. c #846D72",
-"[. c #856D76",
-"{. c #827174",
-"}. c #8A7374",
-"|. c #82727B",
-" X c #8C7679",
-".X c #857B7D",
-"XX c #8B797D",
-"oX c #9C7B75",
-"OX c #947577",
-"+X c #916871",
-"@X c #A3847F",
-"#X c #7B7789",
-"$X c #7E7B94",
-"%X c #777FA3",
-"&X c #837683",
-"*X c #8D7B82",
-"=X c #887C87",
-"-X c #927E88",
-";X c #827D90",
-":X c #8D828B",
-">X c #91848A",
-",X c #96888B",
-"<X c #8A8595",
-"1X c #958C93",
-"2X c #9E9199",
-"3X c #A58A88",
-"4X c #A79496",
-"5X c #B39593",
-"6X c #9B91A3",
-"7X c #8E89A6",
-"8X c #A79AA2",
-"9X c #A49EB8",
-"0X c #BEA5A5",
-"qX c #B3ACB9",
-"wX c #B9B1B9",
-"eX c #ACA2AD",
-"rX c #C2AAAA",
-"tX c #B1ABC1",
-"yX c #CEC7D7",
-/* pixels */
-"                            R.P.A.B.Z.g.1.1.5.g.Z.                                  ",
-"                    #X#X#XI.S.A.Z.g.j.d.s.g.1.g.h.B.Y.>X                            ",
-"                <X<X=X=XS.N.Z.D.C.x.n.F.D.J.K.J.J.K.T.&X=X<X$X                      ",
-"            9X6X6X:X&X&X&X&X X].^.~.~.~.^.}. X X XXX-X-X-X>X2X9X                    ",
-"        %XtXyXtXeX8X8X4X1X-X}.).^.^.^.^.}. X X X}. XXX*X*X*X-X8XqX7X                ",
-"      R.6XwXwX8X1X>X*XXX{.^.F.F.F.F.F.F.^.^.^.^.^.].{.|.|.|.|.|.4XqX;X              ",
-"    V.I.>X>X:X.X|.Y.K.D.D.x.x.j.x.x.x.x.F.F.D.D.J.K.Y.T.|.T.H.H.{.2X1XP.            ",
-"    V.N.L.I.L.A.Z.j.h.j.j.f.d.w.;.8.8.w.s.d.k.j.C.J.K.Y.T.U.H.H.H.|..XS.            ",
-"  2.4.4.N.g.5.1.1.;.;.s.d.f.d.d.d.s.s.d.f.x.x.e.6.=.=.;.s.g.h.B.B.D.Y.A.M.          ",
-"  ,.#.#.#.#.-.-.#.%.W W T T ^ &.=.=.%.%.=.&.&.8.8.8.8.;.;.w.d.C.B.j.j.j.<.          ",
-"  o.$.E E O.#.#.#.@.O.W _ ' &.=.=.w.f.f.j.e.=.6.=.=.;.;.w.>.;.*.=.w.s.d.s.2.        ",
-"  B o.R N E @.#.@.W Y U F A F F T U J H [ ] ! ! T T Y W E B b n N R -.s.g.5.        ",
-"  b R R N M U U U U R W R R R U U =.%.Q H ` ] ] U U U l j c ' d.f.w.;.1.s.g..X      ",
-"  v m R M z s - * > : 3 s l R %.Y ' f.w./ ` } ] T Q G G H `  .t.y.v.v.e.g.h.x.)     ",
-"  v m R M l 3 * * < < 1 1 9 k ! ! H ` ] ! P ~ ~ P P ! ! H A w w Z P | q.e.D.J.z.    ",
-"1X5.m E R l h k A e w w e e C Z D Z D K D D L P I ~ H k < O * 3 s 4 w P ` e.l.u.    ",
-"^ E m o.O.M k Z e 0 5 , , 2 0 e C C V V Z Z D P I D w + o o : v #.@.h A H [ d.) X.  ",
-"6 ( B #.@.z e 0 , o . X X X o , 2 0 r w e H / } ` Z 2 < 4 > o - -.A.;.w.s.s.s.e.u.~.",
-"A G U ,.O.h 0 , . - & . # * - o o + , 2 Q D.}._.z.6.z 4 s s s l T ] 6.b.G.D.h.b./.~.",
-"A r x <.E 9 < o . u 3 o * 3 s * o O 1 q 8.XX3X3X`.~.j.;.U h q k J ] l._.^.n.f.b.^.{.",
-"! C A <.E l l j u 3 9 9 9 9 4 4 4 3 q J e.,X4XrX@XQ.n.l.l.e.=.6.q.y.Q.W._.n.m.b.b.  ",
-"  { ^ #.B N @.*.T k q q q q q j q 4 k G } Q.3XrX5XQ.E.q.|  . . .| q.y.v.~.).G.n.l.  ",
-"  8.8.@.M U U U Q ! T H F A F z q 2 w H } z.3X0XrXoXQ.t.` L L L ~ ..q.l.W.).G.m.l.  ",
-"    W W Y Y _ *.&.=.6.=.[ ! F w 2 5 k ] 9.z.oX5X0X5X'.`.y.P D K K P ` [ 6.e.f.f.f.s.",
-"      W R _ %.8.r.6.e.e.r.[ G 0 5 x ] 6.7.0.y.Q.(.oX'.`.(.u.~ P P H T ^ %.&.=.=.w.>.",
-"        W Q J F Q T ' [ [ / J w q T / ! I I I .. .t.7.} 7.u.l.l.l.f.>.-.-.*.w.w.;.@.",
-"        W U j k F U U T ' { { T Q A Z k w Z D D D q * 3 F f.G.G.b.j.-.m l M W =.@.b ",
-"          c z c k h 9 h k U ' ] %.T 4 . X . O + o o 3 N -.3.3.3.5.@.M f l c _ *.m l ",
-"          c c j 9 3 s s l N E @.@.E f > & o @ > * 9 Y *.3.g.g.s.p.s.s.-.E E *.O.f   ",
-"            j j 9 9 s - > 3 s l l c M M M z c W ^ %.:.8.k.J.].^.K.D.h.;._ E E v d   ",
-"            8 3 s 9 s s s s f l n U +.8.d.a.c.c.!.!.!.+XOXOX].K.D.Z.h.-.B f d a     ",
-"              8 : 3 3 8 s l M N R W +.:.i.i.a.a.a.i.8.8.8.;.s.j.b.b.-.m a t t t     ",
-"                a & > : : s l z z z l j h h l l l l z c U _ >.f.b.3.m : ; ; t       ",
-"                  7 & & & & : s j z k h 9 9 9 3 9 s l c T [ r.f.s.@.d : : i i       ",
-"                    i = % . # & : 8 h j j h h s s h z U T ' [ w.*.b i t i p d       ",
-"                      g > # . # % & 3 s s h j j z M Y ' *.=.*.E m d : t ; i         ",
-"                        a = # # % % & : s f l M M M Y E E E m f t : : y t a         ",
-"                          a & % % % # % & : a s s l l l f d : : ; = = ; t           ",
-"                            i = % % % = % % & & > 3 3 3 : ; & = ; = = t             ",
-"                              : & % % & % % % . & & & & ; = ; = ; t t               ",
-"                                  & % % % % % % % & = = = = t i i                   ",
-"                                          & = = & = ; ; ; t i                       "
-};
