+++ img/purplestone.xpm
@@ -1,0 +1,90 @@
+/* XPM */
+static char *e1316a6c007747d3a2ef8a7931a86de2[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 20 1 ",
+"  c #382C14",
+". c gray22",
+"X c #403018",
+"o c #483818",
+"O c #543C1C",
+"+ c #5C4020",
+"@ c #6C4824",
+"# c #744C28",
+"$ c #400040",
+"% c #480048",
+"& c #500050",
+"* c #580058",
+"= c #6C0070",
+"- c #484848",
+"; c gray33",
+": c #646464",
+"> c #80502C",
+", c #840084",
+"< c #9C009C",
+"1 c #B400B4",
+/* pixels */
+"===$$$$$*=,*=$$$$$$$======*$$$$$$$$$*=,,=**$$$$$$$$$$$$$********",
+"+,,=$$$*=,==**$$$$$=,<@@@+=*$$$$$$&*=,,,,=,=$$$$$$$$$***=,,@@@@@",
+"@+,,,=$*==*===*$$&*,<@@++++**&$$&&*&=,,,=,==*$$$$$$$***=,,@@@@@@",
+"+++,=*****==**$%&&*,<@+,==+=*&&&**&*=,=,==***&&%&&%&&**=,@@@@OO*",
+",+o=,=******&%%&&**=,,=*=,@=&*+++,=**=,=**++@+++OOoo&&&=,@@@X*O=",
+",+o,,==*==,,,,,==***=,=,,=O&*+@@@+,=**=oo+@@@@@+++o+o&&=,=@+X=O,",
+"=,+=,=,=,,,,@==,,==**=*,==X*+@@OO+,=*Xo@@#@@@+OOXOXX+o&*=,@+X,O=",
+",,+,===,+#,@+=*===,=****=*o*@+O**O=*XO@#####@oOo$O**+Xo&*,@+X,O*",
+"=,+=*=,+##@+O=****==*=**==+=++*,=o,=X@##>>##@o$o*O==O*o&&=+OX=O&",
+"*=+*=,,##@++O=**=*****====O*++=,,X=X@Oo>>##@+X*X*O,,O=o*&*+OX,O&",
+"**X==,<#@++Oo=*==+++Oo***=o&o*,==X&XO@=>#@@+OX***,,OO=oX*&*OX=*&",
+"**X=,,<#++ooo=*=+@@@@+=****&o=,*=X%XOo=>@@+OOOX,*,OOOOoX=*&OX&*&",
+"**X=,<<=+oX=X*=*@@+=+@,=*&%%X&&&&%%*Oo,@@+O*=OOX,,OooOoX==*OX&&%",
+"**X=<,@@@+X&X&&+@+==,+,=&$$%&&*&&&*&Oo,@+O*=,=OX+@oXX,oX=,*&&&%%",
+"***=,@@+==== %&+@+*,=+,=&%%&**=,,=&&oo,@*O=,,,@+@OX===oX=,=*&&%%",
+"***=,@+=,<<, $&+@+*,*O=*%*ooo*====&%o*,@,O*oO@@@OX=,=*=oX==&%%%%",
+"***=,@=,<,== $&@,O*=*o&&&+@@++o*=*%%o*,@,Oo@@@@@@OX==*,ooX=&%%%%",
+"***==@=<,**$ $&@,o*=*X&*@>>>>@+**=%%X=,*,O@@@@OOOOX=,===oXX&%%&*",
+"***==@=<*=,* %&,,=,=*&&,>@++++@O**&%X==**O@@OO O OX==**,=X&$&&&*",
+"****==,=*,=* &*=,,=*&&*,+=1<,<+OoX*&X=,*OO@Oo *O*O =*=*,=X&%&**=",
+"===**=,**==,=*=,==*&&*,@11<,,O,OOo**X==**O@O *=*=O =*=*==o&&====",
+"@,,=*==*=**=,=***&&&=,@#,<,@OOXOXXX*o=***O,O ==** **=*&=*o*=,##@",
+"@@+,=****=,**$$&**==,@#@@@@+X*=X*XXoOo**=O,o ***==*=*&&&oO,,##@@",
+"++@o=**=,,,=*$$++X*=,@#=,@OO*=,*XXO+@+**==,X=****,=**&%$o+@@##@@",
+"+OOo=,==,,,=&$O+Oo*=,@#=,@OX*,*XXoooOOO===,X,==,,=,=%$&&+@##,,,+",
+"**oXX==,,,=*$&@=*o*=,@#@@OXX=,OooXXOOoOOX,===,,==*=$$%*=@##<,===",
+"***oo=*===*o@@@*=XX=,@@=**==,,OoXo**=OOoXX=,==,=*&%$$%=,##<<,*=*",
+",,,OO*,=**oOO+***X&=,,@=*=@@@OXXoOO,<O=OoXX=**=*&&%%%*,#@####@**",
+"*==++*=*&&o==+==,=&*=,@,<<<<<OX==oO,,O=OOoX*&&*&+++**=,@#+++##@=",
+"*=,@+,*&&&&*=o,,=*%&*,,<<<,,,OX*<,X*=O=OOoX*&&&+++**=,<@+,11+@@,",
+"=,=++=*=***==XoooX%%*,,<<,,=+OX=,,X**O*OOoX=*o&++,,,,,<@,1,,<<@+",
+"=,=OO,=,=*=,<<,,=OXX=,<,+@@OXO=<==X*===OOoX=*&&+,,,==,<@,1==<<++",
+"*,=oo,*=**,11<+++,o&,,,+@+OoO=<,=,*=,,,OOoX=*%&o=,,=*,,@,<=*,,OO",
+"*=,OO===*=<1<+++++=*,,,@+X%oOo<,*===,,,OOoX=*%&o===*=,,@,<=*=ooo",
+"*=,@+=,,*,<<++oo=o,*,,=@o&**oo,=%*=,,=*OXoX=*%&o**=**=,+=,=&=ooo",
+",*,+o=,=*,,<=+oo,o=*,,*o%*=*oX=*%*=,,=*OX*=*&&&o&**&**=+=,=*&&OO",
+"**,oo==**=,,=ooo=o**=,=o%==*XX&%***=,**OX*=*&**o&&&&&&*O*===**++",
+"**=o=**&&=,,,ooo*o%&*==+$**%XX%%=**,**==X=,====o&&&**&&O$*====@@",
+"**,=*&&&&*==,=o*&o%%&**o$=*%XX%**%**==*=X*==,,,o&*====*O&$&**&@@",
+"**=***&&&&*==*o*&&&&%&&X%**,=X$%%**=,,*=*X*=,,=o*=,,,,=**%$$$$++",
+"***====*=*&*=,,,,=&*&&*X%%%=* $%%*=,,=***X*,=**oo++,,,,==*%%$$Oo",
+"**=,@@,=,=*=,,,,,,,*=** %%%%% $%**=,=**=*&*&*,o++@@++,==,=*%%$oo",
+"&*=@++O==*==,,@@OO==,=& &%%%%X$*==**&&&*&%&**=,+@+O*O=,=,,=*%%XX",
+"&*,+==O**=,,,@@@OO**=*%X&&&&&o&****=**&&%%&*=,<@@Oo$o=,,,=,=*&X ",
+"&*,O=,O,=,,@@@++Oo***&$o*OOOO+,,***===**&&&*=,<O,o*$**=,=*=*=* $",
+"&*=O=*O==,,+,@**OX*=*&$oO@@++@<<,**=,,==***=,@<O,o$$*=,,=**=,= $",
+"&**o=,,=,,=O*,&**X=,=&%O@>#@+@,1<@o=,,=,=*=,@@,o=*$**o+<,===,= $",
+"%&*o=,,=,,=*,*&=,<X,=&&+>##@@@=<,+@o,=,==*=@@@,o=**oo+@#,=,===**",
+"&%&X===*=,=**&==<=X=***@#@@+@@@,<,++=,=,==@@@@@@*oo++@#+==,==,=X",
+"&%&X&****=***=*==X=,=*@@+O@O+@@@,<,+o====,@@++++@,,###+o**===,=X",
+"%$% %&&&**==&*&%$X*=**,@O++oO+++=,,+o&**=@@++%&&*==+o*+o*=,,=,=X",
+"$$$ %$%%&&***==%%X&*o@@@@+o+OOoo+@,+o%&**@+=O&**=,,+o=+o=,,=*=**",
+"%$$$$%&&*******&&X o@OOo+o++ooooo++++o%&*O,=o*o++oX+,,=o,,O+o**X",
+"&$$ %&**==,,,,,=*Xo@OOX oo+oo%**==OO++oX*o=,o=+OOoX+,==o=O+++o*X",
+"%$%%&*=,,,@@@@@,=*o+OX% X%ooX*,==*ooooOo%X*=o,=o%oX+=**X,==OOo&&",
+"%%%%&=,<<+@OOOO@,*o+X%% X%ooX=,,,,XX  Oo&X%*X,=o%XX**=*X=*=ooo&&",
+"%%%%*=,<<@*===OO=*o+X%%%X%%*X*=,,, X %oX**&%X==X%oo**,*X***ooo&&",
+"&%%&*=,<*O=,=,=O=* ,,**,X*,*X*=,==  %&X &*&%X&&X&OOX***X*=**+OO*",
+"&%%&*=,OOXOO,=,o== *,,**X***X=*=*%% &=X %=%&X*&&*+=+X**X=**+>@@O",
+"&%%&O@@@=,=====o=, *****XXXXX==***% =*X %&&*X==*,O,=XX*X**+>@++,",
+"X&&=@@++@==,,XXXXXX**,*XooooooXXXXX X*X %=*=,,,=*XXXXXXXXX@>+=XX",
+"XXX,@&=,O,*=XooooooooXXX+,++++OOOoooXXX =,==,,,oooOOOO+++@>+o,=X",
+"XXXX+=,*oXXXo++@@+++ooooo+@#>#@@+++++OOooooXXXXoooO+@@@@@@+oXXXX",
+":-.XXXXXXXXXXXXXXX.-::;.XXXXXXXXXXXXXXXXXXX.-;;-.XXXXXXXXXXXX.-:"
+};
