+++ img/tonneau.xpm
@@ -1,0 +1,97 @@
+/* XPM */
+static char *_573127374323[] = {
+/* columns rows colors chars-per-pixel */
+"64 64 27 1 ",
+"  c black",
+". c gray11",
+"X c #231C0A",
+"o c #312611",
+"O c #382A15",
+"+ c #3F3115",
+"@ c gray15",
+"# c #313131",
+"$ c #3F3F3F",
+"% c #493418",
+"& c #50381C",
+"* c #5E3F1F",
+"= c #654223",
+"- c #704626",
+"; c #774D2A",
+": c #7E502D",
+"> c #494949",
+", c gray34",
+"< c #626262",
+"1 c #6C6C6C",
+"2 c gray48",
+"3 c #885431",
+"4 c #8C5734",
+"5 c #935B38",
+"6 c #9D623B",
+"7 c #A4693F",
+"8 c gray52",
+/* pixels */
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                                                                ",
+"                       <<<<<<<<<1111112222                      ",
+"                     1<>>>>>>>>>>>,,,,,<<<21                    ",
+"                    12>#....@@@@@@@##$$$$><><                   ",
+"                    112882211111<<<<,,,,>>$##                   ",
+"                   33112211<<<,,,,>>>$$$####++                  ",
+"                   ;4:33;---====***&&&++++++&+                  ",
+"                  3;446737664:5543:=*-==%=*++&+                 ",
+"                  ;4;6745766:;5543:=*--=*&=*+&+                 ",
+"                  ;4;6737666::5543:-&---=%=*+&$                 ",
+"                 >,12<737665:5543:;-&---=%$$$$#                 ",
+"                 ><1122821111<<<<<,,,,,>>>>####+                ",
+"                 >111112<<<<,,,,>>>>>>$$$$#####+                ",
+"                 ;=>,<11<<<<,,,,>>>>>>$$$$$###&+                ",
+"                 ;4;644&3:;-+===****X&++++%&&+&+                ",
+"                 ;4;67747666:5433:;-&---=*%&&+&+                ",
+"                 ;4;67747666:5433:;-&---=*%&&+&+                ",
+"                 ;4;67747666;5433:;-&---=*%&&+&+                ",
+"                 ;4;67747666;5433:;-&---=*%&&+&+                ",
+"                 ;4;67747666;5433:;-&---=*%&&+&+                ",
+"                 ;4;67747666;5433:;-&---=*%&&+&+                ",
+"                 14;67737666-5433:;-&---=*%&&+$$                ",
+"                 <2221737666-5433:;-&---=*%$$$$#                ",
+"                 ,11128211111<<<<<,,,,,>>>>$###                 ",
+"                 >,<1121<<<<,,,,>>>>>>$$$$$####O                ",
+"                  =>,,,<<<<<,,,,>>>>>>$$$$$OOO%&                ",
+"                  ;;------**o**&&&&&X&&&&OO&+&+&                ",
+"                  ;4;5737553=3::::;-&--=*+&&+&+                 ",
+"                   ;446;5553=;::::;*&--=*+&&+&+                 ",
+"                   <4;44;553=-:::;-&&--=&+&+&+                  ",
+"                   <2824;333-=;;;;-&*-=*+&+++#                  ",
+"                    1112222111<<,,,>>>$$$##@.                   ",
+"                     1<11<<<,,,,>>>>$$$###@@                    ",
+"                       >>>>>$$$$$####@@@@@                      "
+};
